Tighten up nvpair naming
The List enum variant and the List type having the same name was
confusing, and then as I got into changing that it occured to me that
exposing the 'nv' prefix wasn't super helpful either, because really
we're just dealing in lists. I might reneg on that last part, we'll see.