Lift handle into a shareable internal type
We're going to be creating more objects representing ZFS concepts
(pools, datasets, etc), each of which will need to make their own calls
down into ZFS to fill in the details. So this sets up for the handle
(ioc handle and caches) to be shared by all those objects so that they
can have indepnent lifetimes.