Decouple the ioc handle from top handle mutability
Mutability here is less about the data, more about whether or not we're
changing things on disk. Even for "read only" operations we need to
mutate the ioc handle, thus RefCell.

(maybe the ioc handle should get similar treatment but I'm not sure
about this one yet so I'll wait to see how it feels before pushing on it
further).