Remove internal caching
A lot more cloning, but, these are very small items, and this library is
mostly going to be for interactive or short-lived programs, and this
takes care of staleness issues that I had been avoiding thining about.

Justification is roughly that you wouldn't think twice about copying and
refetching JSON, and the performance characteristics we want are about
the same, so why try harder?